requirements.txt
aiomcache==0.6.0
aioredis==1.3.1
async-timeout==3.0.1
asyncio==3.4.3
asyncio-redis==0.16.0
asyncpg==0.21.0
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
et-xmlfile==1.0.1
hiredis==1.1.0
idna==2.10
jdcal==1.4.1
multidict==5.0.0
numpy==1.19.2
oauthlib==3.1.0
openpyxl==3.0.5
outcome==1.0.1
pandas==1.1.3
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pylibmc==1.6.1
pytz==2020.1
redis==3.5.3
Represent==1.6.0
requests==2.24.0
requests-oauthlib==1.3.0
rethinkdb==2.4.7
six==1.15.0
SQLAlchemy==1.3.20
sqlalchemy-aio==0.16.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.25.11
XlsxWriter==1.3.7
xlwt==1.3.0
